requirements.txt
spacy
fuzzywuzzy
rapidfuzz
numpy
textblob
nltk
imdbpy